JPMS support
Closes #133
