+++ api/src/main/resources/META-INF/services/io.vertx.core.spi.VertxServiceProvider
