+++ log4j2/src/test/log4j2/context/resources/META-INF/services/org.apache.logging.log4j.core.util.ContextDataProvider
@@ -1,0 +1,17 @@
+#
+# Copyright 2021 Red Hat, Inc.
+#
+# Red Hat licenses this file to you under the Apache License, version 2.0
+# (the "License"); you may not use this file except in compliance with the
+# License.  You may obtain a copy of the License at:
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+
+io.reactiverse.contextual.logging.log4j.impl.VertxContextDataProvider
