+++ logback/src/test/java/io/reactiverse/contextual/logging/logback/tests/LogbackContextualLoggingIT.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2024 Red Hat, Inc.
+ *
+ * Red Hat licenses this file to you under the Apache License, version 2.0
+ * (the "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at:
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.reactiverse.contextual.logging.logback.tests;
+
+import io.reactiverse.contextual.logging.tests.ContextualLoggingIT;
+
+public class LogbackContextualLoggingIT extends ContextualLoggingIT {
+}
