--- src/test/jul/logging.properties
@@ -1,21 +1,0 @@
-#
-# Copyright 2022 Red Hat, Inc.
-#
-# Red Hat licenses this file to you under the Apache License, version 2.0
-# (the "License"); you may not use this file except in compliance with the
-# License.  You may obtain a copy of the License at:
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
-# License for the specific language governing permissions and limitations
-# under the License.
-#
-handlers=java.util.logging.ConsoleHandler
-java.util.logging.FileHandler.pattern=target/ContextualLoggingIT.log
-java.util.logging.FileHandler.formatter=io.reactiverse.contextual.logging.JULContextualDataFormatter
-io.reactiverse.contextual.logging.JULContextualDataFormatter.format=%{requestId:-foobar}$s ### %5$s%n
-io.reactiverse.contextual.logging.handlers=java.util.logging.FileHandler
-io.reactiverse.contextual.logging.useParentHandlers=false
