--- src/test/logback/logback.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-  ~ Copyright 2022 Red Hat, Inc.
-  ~
-  ~ Red Hat licenses this file to you under the Apache License, version 2.0
-  ~ (the "License"); you may not use this file except in compliance with the
-  ~ License.  You may obtain a copy of the License at:
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
-  ~ License for the specific language governing permissions and limitations
-  ~ under the License.
-  -->
-<!DOCTYPE configuration>
-
-<configuration>
-
-  <import class="ch.qos.logback.classic.encoder.PatternLayoutEncoder"/>
-  <import class="ch.qos.logback.core.ConsoleAppender"/>
-  <import class="ch.qos.logback.core.FileAppender"/>
-  <import class="ch.qos.logback.core.status.NopStatusListener"/>
-
-  <statusListener class="NopStatusListener"/>
-
-  <conversionRule conversionWord="vcl" converterClass="io.reactiverse.contextual.logging.LogbackConverter"/>
-
-  <appender name="STDOUT" class="ConsoleAppender"/>
-
-  <appender name="FILE" class="FileAppender">
-    <file>target/ContextualLoggingIT.log</file>
-    <append>false</append>
-    <encoder class="PatternLayoutEncoder">
-      <pattern>%vcl{requestId:-foobar} ### %msg%n</pattern>
-    </encoder>
-  </appender>
-
-  <logger name="io.reactiverse.contextual.logging" level="INFO" additivity="false">
-    <appender-ref ref="FILE"/>
-  </logger>
-
-  <root level="INFO">
-    <appender-ref ref="STDOUT"/>
-  </root>
-
-</configuration>
